+++ src/main.js
@@ -8,6 +8,7 @@ define(
         var config = {};
         var u = require('underscore');
         var util = require('./util');
+        var Deferred = require('er/Deferred');
 
         /**
          * 初始化API请求器
@@ -18,6 +19,11 @@ define(
             config.api = util.genRequesters(config.api);
         }
 
+        /**
+         * 初始化ER 配置
+         *
+         * @ignore
+         */
         function initErConfigs() {
             var erConfig = require('er/config');
             erConfig.indexURL = config.index;
@@ -27,20 +33,36 @@ define(
         /**
          * 初始化系统启动
          *
+         * @param {Array} [extra] 额外的请求发送器
+         * @return {er.Promise} 处理完毕的Promise
          * @ignore
          */
-        function loadData() {
-            var Deferred = require('er/Deferred');
-            return Deferred.all(
-                Deferred.when(config.api.user()),
-                Deferred.when(config.api.constants())
+        function loadData(extra) {
+            extra = extra ? u.map(extra, function (api) {
+                if (typeof api === 'string') {
+                    return util.genRequesters(api);
+                }
+                else {
+                    return api;
+                }
+            }) : [];
+
+            var requests = [config.api.user, config.api.constants].concat(extra || []);
+
+            return Deferred.all.apply(
+                Deferred,
+                u.map(requests, function (requester) {
+                    return Deferred.when(requester());
+                })
             );
         }
 
         /**
          * 默认读取用户信息和系统常量后初始化对应模块
          *
-         * @ignore
+         * @param {Object} session 用户信息
+         * @param {Object} constants 服务器端常量
+         * @return {er.Promise} 处理完毕的Promise
          */
         function initData(session, constants) {
             // 初始化用户信息
@@ -51,6 +73,15 @@ define(
             var consts = require('./system/constants');
             var localConstants = require('common/constants');
             consts.init(u.extend(localConstants, constants));
+
+            // 返回其余请求结果
+            var extra = [].slice.call(arguments).slice(2);
+            return Deferred.all.apply(
+                Deferred,
+                u.map(extra, function (result) {
+                    return Deferred.resolved(result);
+                })
+            );
         }
 
         /**
@@ -69,9 +100,13 @@ define(
         /**
          * RIA启动入口
          *
+         * @param {Object} riaConfig RIA配置
+         * @param {Array} requesters 初始化数据需要的请求发送器
+         * @param {Function} callback 初始化请求返回后的回调函数
+         * @return {er.Promise} 处理完毕的Promise
          * @ignore
          */
-        function start(riaConfig) {
+        function start(riaConfig, requesters, callback) {
 
             config = riaConfig;
 
@@ -93,8 +128,9 @@ define(
             initApiConfig();
 
             // 读取必要信息后初始化系统
-            return loadData()
+            return loadData(requesters)
                 .then(initData)
+                .then(callback)
                 .then(init);
         }