+++ src/ui/loading.js
@@ -1,3 +1,8 @@
+/**
+ * @file 加载提示框，基于esui.Dialog实现
+ * @author Justineo(justice360@gmail.com)
+ */
+
 define(function (require) {
     var u = require('underscore');
 
@@ -38,7 +43,9 @@ define(function (require) {
         properties = u.extend(properties, options);
         globalLoading.setProperties(properties);
         globalLoading.show();
+        globalLoading.main.style.zIndex = parseInt(globalLoading.main.style.zIndex, 10) + 1;
         loadingCount++;
+        loadingTimer && clearTimeout(loadingTimer);
         return globalLoading;
     }
     function hideLoading() {