+++ src/ui/navigator.js
@@ -1,10 +1,8 @@
 /**
- * Copyright 2014 Baidu Inc. All rights reserved.
- * 
  * @file Navigator
  * @author chestnutchen(chenli11@baidu.com)
- * @date $DATE$
  */
+
 define(function (require) {
     var u = require('underscore');
     var locator = require('er/locator');
@@ -13,18 +11,17 @@ define(function (require) {
     var URL = require('er/URL');
 
     /**
-     * @class Navigator
-     *
      * 导航单例
      * 使用当前er action path来进行匹配确定是否高亮
      * include规则进行匹配
      * exclude规则进行过滤
      *
-     * @usage
+     * 使用方法：
      * 在html中放一个div或者ul元素，<ul id="nav" class="nav"></ul>
      * 初始化er前
      * req('bat-ria/ui/navigator').init('nav', globalConfig.nav);
      *
+     * @constructor
      * @singleton
      */
     function Navigator() {}
@@ -36,21 +33,20 @@ define(function (require) {
     /**
      * 初始化导航
      *
-     * @param {String} domId  dom元素id
+     * @param {string} domId  dom元素id
      * @param {Array} config  配置数组
      *
-     * @cfg {String} [config.navId]  主导航dom元素id
-     * @cfg {String} [config.text]  导航文本
-     * @cfg {String} [config.url]  er.url，不包含参数
-     * @cfg {String} [config.externalUrl]  外部路径，优先跳转
+     * @cfg {string} config.navId  主导航dom元素id
+     * @cfg {string} config.text  导航文本
+     * @cfg {string} [config.url]  er.url，不包含参数
+     * @cfg {string} [config.externalUrl]  外部路径，优先跳转
      * @cfg {Array} [config.include]  表示是否需要高亮该导航
      *      匹配目标均为`er.action.url`，不包含`~`的参数
      * @cfg {Array} [config.exclude]  和`include`相反的配置
      *      这里的逻辑`exclude`优先级比`include`高，但是也不要两个都配同个规则吧。。
-     * @cfg {Array} [config.tabs]  子导航，结构和一级导航config中每一项保持一致
+     * @cfg {Array} config.tabs  导航项，子导航结构和一级导航config中每一项保持一致
      * @cfg {string} [config.auth]  与er.permission对应的权限控制
-     *      TODO：二级导航暂未支持权限
-     * 
+     *
      * @sample:
      * nav: {
      *      navId: 'nav',
@@ -112,7 +108,7 @@ define(function (require) {
      */
     function toggleNav(type) {
         this.main.style.cssText = 'display:' + type + ';';
-        u.some(this.subNavs, function (subNav, index) {
+        u.some(this.subNavs, function (subNav) {
             if (u.isObject(subNav)) {
                 if (lib.hasClass(subNav.nav, 'nav-sub-item-current')) {
                     subNav.nav.cssText = 'display:' + type + ';';
@@ -134,6 +130,8 @@ define(function (require) {
      * locator redirect的时候处理navigator的主要逻辑
      * 先将对应一级导航高亮
      * 如果该导航下有二级导航，创建或者显示出来，并高亮二级导航对应元素
+     *
+     * @param {Event} e 事件对象
      */
     Navigator.prototype.handleRedirect = function (e) {
         var me = this;
@@ -172,10 +170,10 @@ define(function (require) {
     /**
      * 高亮当前导航元素
      *
-     * @navItems {object} 缓存导航的集合，传入来记录当前高亮的索引
-     * @element {object} 要高亮的元素
-     * @index {string} 要高亮的导航索引
-     * @className {string} 通过添加和移除current className来更换高亮样式
+     * @param {Object} navItems 缓存导航的集合，传入来记录当前高亮的索引
+     * @param {Object} element 要高亮的元素
+     * @param {string} index 要高亮的导航索引
+     * @param {string} className 通过添加和移除current className来更换高亮样式
      */
     function activateNavElement(navItems, element, index, className) {
         if (!u.isNumber(navItems.activeIndex)) {
@@ -192,9 +190,9 @@ define(function (require) {
     /**
      * 展示或隐藏二级导航
      *
-     * @navItems {object} 缓存导航的集合，传入来记录当前高亮的索引
-     * @element {object} 要高亮的元素，不传表示隐藏所有
-     * @index {string} 要高亮的导航索引
+     * @param {Object} navItems 缓存导航的集合，传入来记录当前高亮的索引
+     * @param {Object} [element] 要展示的子tab，不传表示隐藏所有
+     * @param {string} [index] 要展示的导航索引
      */
     function toggleSubNav(navItems, element, index) {
         var className = 'nav-sub-current';
@@ -218,23 +216,63 @@ define(function (require) {
 
     /**
      * 创建nav元素
-     * 
-     * @config {object} globalconfig
-     * @navItems {object} 缓存nav元素的对象
-     * @nav {object} nav父元素
-     * @isSub {string} '' 或 'sub-'，主导航或二级子导航 
+     *
+     * @param {Object} config globalconfig
+     * @param {Object} navItems 缓存nav元素的对象
+     * @param {Object} nav nav父元素
+     * @param {string} [isSub] {''|'sub-'}，主导航或二级子导航
      */
     function createNavElements(config, navItems, nav, isSub) {
-        var isSub = isSub || '';
+        isSub = isSub || '';
         u.each(config, function (item, index) {
             if (!item.auth || permission.isAllow(item.auth)) {
-                var url = item.externalUrl || ('#' + item.url);
+                var internalUrl = fixErUrl(item.url || '');
+                var url = item.externalUrl || internalUrl;
+                var target = (item.externalUrl ? ' target="_blank"' : '');
                 var element = document.createElement('li');
                 var separate = '';
 
+                // 处理父tab和子tab权限耦合的部分
+                if (item.tabs) {
+                    // 看看默认要跳转的地址所在的子tab有没有权限，没有的话得找备胎
+                    var hasDefaultTabAuth = u.every(item.tabs, function (subItem) {
+                        if (isActive(item.url.replace(/^#/, ''), subItem)) {
+                            if (!subItem.auth || permission.isAllow(subItem.auth)) {
+                                return true;
+                            }
+                            else {
+                                return false;
+                            }
+                        }
+                    });
+
+                    // 如果没有权限就得找到第一个有权限的子tab，把父tab跳转链接改成这个
+                    if (!hasDefaultTabAuth) {
+                        // 做最坏的打算，所有不外跳的子tab都没权限，那这个父tab只能直接往外跳了
+                        var externalUrlBackUp = '';
+                        var isInternalUrlFound = u.some(item.tabs, function (subItem) {
+                            if (!subItem.auth || permission.isAllow(subItem.auth)) {
+                                if (subItem.url) {
+                                    url = fixErUrl(subItem.url);
+                                    return true;
+                                }
+                                else if (subItem.externalUrl && !externalUrlBackUp) {
+                                    externalUrlBackUp = subItem.externalUrl;
+                                }
+                            }
+                        });
+
+                        if (!isInternalUrlFound) {
+                            // 如果找不到内跳和外跳的备胎
+                            // 就说明所有子tab都没权限，那就用原来父tab配的url就可以了
+                            url = externalUrlBackUp || url;
+                        }
+                    }
+                }
+
                 element.className = 'nav-' + isSub + 'item';
-                element.innerHTML = '<a href="' + url + '">'
-                                        + '<span>' + u.escape(item.text) + '</span>'
+                element.innerHTML = '<a href="' + url + '"' + target + '>'
+                                        + '<span>' + item.text + '</span>'
                                     + '</a>';
 
                 nav.appendChild(element);
@@ -253,11 +291,11 @@ define(function (require) {
     /**
      * 创建二级nav元素，如果存在，直接展示
      *
-     * @config {object} config.tabs，某个主导航的二级导航配置
-     * @navItems {object} 缓存一级导航的对象，用来计算位置
-     * @subNavs {object} 缓存二级导航的对象
-     * @parentNav {object} 一级导航容器
-     * @index {string} 一级导航navElement的index
+     * @param {Object} config config.tabs，某个主导航的二级导航配置
+     * @param {Object} navItems 缓存一级导航的对象，用来计算位置
+     * @param {Object} subNavs 缓存二级导航的对象
+     * @param {Object} main 一级导航容器
+     * @param {string} index 一级导航navElement的index
      */
     function createOrShowSubNav(config, navItems, subNavs, main, index) {
         var ul = subNavs[index].nav;
@@ -294,8 +332,10 @@ define(function (require) {
     /**
      * 验证url是否匹配
      *
-     * @url {string} 当前页面的er path
-     * @patterns {RegExp | string} 验证的表达式或path
+     * @param {string} url 某个path
+     * @param {RegExp | string} patterns 验证的表达式或path
+     *
+     * @return {boolean}
      */
     function testUrlIn(url, patterns) {
         return u.some(patterns, function(pattern) {
@@ -309,10 +349,25 @@ define(function (require) {
     }
 
     /**
+     * 将传入的path格式化为er path
+     *
+     * @param {string} url 某个path
+     * @return {string}
+     */
+    function fixErUrl(url) {
+        if (!url) {
+            return '';
+        }
+        return (url.indexOf('#') === 0 ? url : ('#' + url));
+    }
+
+    /**
      * 验证url是否命中某个导航元素config的规则
      *
-     * @url {string} 当前页面的er path
-     * @patterns {RegExp | string} 验证的表达式或path
+     * @param {string} url 某个path，不带#
+     * @param {RegExp | string} item 一个tab的配置信息
+     *
+     * @return {boolean}
      */
     function isActive(url, item) {
         return !testUrlIn(url, item.exclude || []) && testUrlIn(url, item.include || []);
@@ -325,7 +380,7 @@ define(function (require) {
         };
     }
 
-    var commonNavigator = new Navigator(); 
+    var commonNavigator = new Navigator();
 
     return {
         init: u.bind(commonNavigator.init, commonNavigator),